backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-pagination==0.12.9
greenlet==2.0.2
h11==0.14.0
idna==3.4
Pillow==10.0.1
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyodbc==4.0.39
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
